Make nap_conf.h private, abstract version strings
