+++ src/board/v2/init.c
@@ -13,6 +13,7 @@
 #include <hal.h>
 
 #include <stdlib.h>
+#include <string.h>
 
 #include <libsbp/flash.h>
 #include <libsbp/sbp.h>
@@ -24,12 +25,17 @@
 #include "peripherals/random.h"
 #include "board/v2/m25_flash.h"
 #include "board/nap/nap_common.h"
-#include "board/nap/nap_conf.h"
+#include "nap/nap_conf.h"
 #include "sbp.h"
 #include "error.h"
 
+#define REQUIRED_NAP_VERSION_STR ("v0.16")
+
 #define SCB_AIRCR_VECTKEY                       (0x5FA << SCB_AIRCR_VECTKEY_Pos)
 #define SCB_AIRCR_SYSRESETREQ                   (1 << 2)
+
+static void nap_version_check(void);
+static s8 compare_version(const char *a, const char *b);
 
 /** Resets the device back into the bootloader. */
 static void reset_callback(u16 sender_id, u8 len, u8 msg[], void* context)
@@ -100,11 +106,28 @@ void init(void)
   fault_handling_setup();
   reset_callback_register();
 
+  nap_version_check();
+  nap_callbacks_setup();
+
   rng_setup();
   srand(random_int());
 
-  nap_callbacks_setup();
   stm_unique_id_callback_register();
+}
+
+static void nap_version_check(void)
+{
+  /* Check we are running a compatible version of the NAP firmware. */
+  char nap_version_string[64] = {0};
+  nap_conf_rd_version_string(nap_version_string);
+  if (compare_version(nap_version_string, REQUIRED_NAP_VERSION_STR) < 0) {
+    while (1) {
+      log_error("NAP firmware version >= %s required, please update!"
+                "(instructions can be found at http://docs.swift-nav.com/)",
+                REQUIRED_NAP_VERSION_STR);
+      chThdSleepSeconds(2);
+    }
+  }
 }
 
 /* Check NAP authentication status. Block and print error message
@@ -131,6 +154,107 @@ void check_nap_auth(void)
 s32 serial_number_get(void)
 {
   return nap_conf_rd_serial_number();
+}
+
+u8 hw_revision_string_get(char *hw_revision_string)
+{
+  strcpy(hw_revision_string, nap_conf_rd_hw_rev_string());
+  return strlen(hw_revision_string);
+}
+
+u8 nap_version_string_get(char *nap_version_string)
+{
+  return nap_conf_rd_version_string(nap_version_string);
+}
+
+/** Compare version strings.
+ * Compares a version of the form 'vX.Y-Z-'. If the first character of the
+ * version is not 'v' then that string will be considered older than any
+ * version string starting with 'v'. Two strings neither starting with 'v' will
+ * compare equal.
+ *
+ * \param a First version string
+ * \param b Second version string
+ * \return `1` if `a > b`, `-1` if `b > a`, `0` if `a == b`
+ */
+static s8 compare_version(const char *a, const char *b)
+{
+  if (a[0] != 'v') {
+    if (b[0] != 'v') {
+      /* Both have old style version strings, no way to compare. */
+      return 0;
+    } else {
+      /* a has an old style version string, so is older. */
+      return -1;
+    }
+  }
+
+  if (b[0] != 'v') {
+    /* b has an old style version string, so is older. */
+    return 1;
+  }
+
+  char buff[5];
+  memset(buff, 0, 5);
+
+  /* Skip initial 'v'. */
+  a++; b++;
+
+  /* Extract the major version numbers. */
+  u8 major_span = strchr(a, '.') - a;
+  strncpy(buff, a, major_span);
+  u8 major_a = atoi(buff);
+  a += major_span + 1;
+  memset(buff, 0, 5);
+
+  major_span = strchr(b, '.') - b;
+  strncpy(buff, b, major_span);
+  u8 major_b = atoi(buff);
+  b += major_span + 1;
+  memset(buff, 0, 5);
+
+  if (major_a != major_b) {
+    return (major_a < major_b) ? -1 : 1;
+  }
+
+  u8 commit_a = 0;
+  u8 commit_b = 0;
+  u8 minor_a, minor_b;
+
+  /* Check if we have a commit number. */
+  if (strchr(a, '-')) {
+    /* Extract the minor version numbers. */
+    u8 minor_span = strchr(a, '-') - a;
+    strncpy(buff, a, minor_span);
+    minor_a = atoi(buff);
+    a += minor_span + 1;
+    memset(buff, 0, 5);
+
+    /* Extract the commit numbers. */
+    commit_a = atoi(a);
+  } else {
+    minor_a = atoi(a);
+  }
+
+  /* Check if we have a commit number. */
+  if (strchr(b, '-')) {
+    /* Extract the minor version numbers. */
+    u8 minor_span = strchr(b, '-') - b;
+    strncpy(buff, b, minor_span);
+    minor_b = atoi(buff);
+    b += minor_span + 1;
+
+    /* Extract the commit numbers. */
+    commit_b = atoi(b);
+  } else {
+    minor_b = atoi(b);
+  }
+
+  if (minor_a != minor_b) {
+    return (minor_a < minor_b) ? -1 : 1;
+  }
+
+  return (commit_a < commit_b) ? -1 : (commit_a > commit_b);
 }
 
 /** Our own basic implementation of sbrk().
