+++ src/board/v3/nap/nap_common.c
@@ -26,8 +26,7 @@
 
 #define PROCESS_PERIOD_ms (1000)
 
-void nap_isr(void *context);
-void nap_unlock(const u8 key[]);
+static void nap_isr(void *context);
 
 static BSEMAPHORE_DECL(nap_exti_sem, TRUE);
 static WORKING_AREA_CCM(wa_nap_exti, 2000);
@@ -80,7 +79,7 @@ u64 nap_timing_count(void)
   return total_count;
 }
 
-void nap_isr(void *context)
+static void nap_isr(void *context)
 {
   (void)context;
   chSysLockFromISR();
@@ -126,44 +125,7 @@ static void nap_exti_thread(void *arg)
   }
 }
 
-u32 nap_conf_rd_version(void)
-{
-  NAP->CONTROL = (NAP->CONTROL & ~((u32)NAP_CONTROL_VERSION_ADDR_Msk));
-  return NAP->VERSION;
-}
-
-u8 nap_conf_rd_version_string(char version_string[])
-{
-  u8 i = 0;
-  u32 reg = 0;
-  u32 ctrl = (NAP->CONTROL & ~((u32)NAP_CONTROL_VERSION_ADDR_Msk));
-
-  do {
-    NAP->CONTROL = ctrl | ((i + NAP_VERSION_STRING_OFFSET) / sizeof(reg) <<
-        NAP_CONTROL_VERSION_ADDR_Pos);
-    reg = NAP->VERSION;
-    memcpy(&version_string[i], &reg, sizeof(reg));
-    i += sizeof(reg);
-  } while (reg && i < NAP_VERSION_STRING_LENGTH);
-  version_string[i] = 0;
-
-  return strlen(version_string);
-}
-
-void nap_rd_dna(u8 dna[])
-{
-  u32 reg = 0;
-  u32 ctrl = (NAP->CONTROL & ~((u32)NAP_CONTROL_VERSION_ADDR_Msk));
-
-  for (u8 i = 0; i < NAP_DNA_LENGTH; i += sizeof(reg)) {
-    NAP->CONTROL = ctrl | ((i + NAP_DNA_OFFSET) / sizeof(reg) <<
-        NAP_CONTROL_VERSION_ADDR_Pos);
-    reg = NAP->VERSION;
-    memcpy(&dna[i], &reg, sizeof(reg));
-  }
-}
-
-void nap_rd_dna_callback(u16 sender_id, u8 len, u8 msg[], void* context)
+static void nap_rd_dna_callback(u16 sender_id, u8 len, u8 msg[], void* context)
 {
   (void)sender_id; (void)len; (void)msg; (void) context;
   sbp_send_msg(SBP_MSG_NAP_DEVICE_DNA_RESP, NAP_DNA_LENGTH, nap_dna);
@@ -175,15 +137,4 @@ void nap_callbacks_setup(void)
 
   sbp_register_cbk(SBP_MSG_NAP_DEVICE_DNA_REQ, &nap_rd_dna_callback,
       &nap_dna_node);
-}
-
-void nap_unlock(const u8 key[])
-{
-  u32 ctrl = (NAP->CONTROL & ~((u32)NAP_CONTROL_KEY_ADDR_Msk |
-        NAP_CONTROL_KEY_BYTE_Msk));
-
-  for (u8 i = 0; i < NAP_KEY_LENGTH; ++i) {
-    NAP->CONTROL = ctrl | ((u32)key[i] << NAP_CONTROL_KEY_BYTE_Pos) |
-        (i << NAP_CONTROL_KEY_ADDR_Pos);
-  }
 }
